+++ RELEASE-NOTES.md
@@ -13,6 +13,7 @@
 * #334 AvroLoader: "string" type without "avro.java.string":"String" is processed incorrectly
 * #335 Stop adding "avro.java.string" property to String field schemas
 * #337 Compacting Parquet columns on load
+* #338 Compacting JDBC columns on load
 
 ## Release 1.0.0-M23